Move `components/Chat.vue` into `components/Chat/index.vue`
Also move `css/chat.css` into `components/Chat/styles.css`
