+++ app/src/renderer/components/Chat/index.vue
@@ -1,0 +1,27 @@
+<script>
+import ThreadSection from './ThreadSection.vue'
+import MessageSection from './MessageSection.vue'
+
+export default {
+  name: 'Chat',
+  components: {
+    ThreadSection,
+    MessageSection
+  },
+  mounted: function () {
+    this.$store.dispatch('iotConnect')
+  },
+  destroyed: function () {
+    this.$store.dispatch('iotDisconnect')
+  }
+}
+</script>
+
+<style src="./styles.css"></style>
+
+<template>
+  <div class="chatapp">
+    <thread-section></thread-section>
+    <message-section></message-section>
+  </div>
+</template>
