+++ app/src/renderer/components/Chat/styles.css
